Service 역할 분리 - SignupSerivce: 회원가입 기능 - StaticFileService: 정적 파일 응답
