+++ src/main/java/service/Service.java
@@ -1,0 +1,9 @@
+package service;
+
+import http.HttpRequest;
+import http.HttpResponse;
+
+public interface Service {
+
+    HttpResponse run(HttpRequest httpRequest);
+}
